🚚 Rename: 마이페이지 관련 디렉토리 통합 (#93)
