+++ src/components/index.tsx
@@ -3,7 +3,7 @@ import Badge from '@/components/common/Badge'
 import Button from '@/components/common/Button'
 import ImageCard from '@/components/common/Card/ImageCard'
 import Input from '@/components/common/Input'
-import MyPageSideBar from '@/components/my-page-side-bar/MyPageSideBar'
+import MyPageSideBar from '@/components/mypage/MyPageSideBar'
 import Header from '@/components/header/Header'
 import { Symbol } from '@/components/common/Symbol'
 import { Logo } from '@/components/common/Logo'
